detector/requirements.txt
fastapi
uvicorn
kafka-python
pydantic
pydantic-settings
